Refactor corners setter in cdll_line.py
The corners setter in the cdll_line.py file was refactored for optimization. The iteration over corners and nodes in the setter has been modified to use enumeration. Additionally, adjustments have been made to the 'set_corner' method to make sure a new array is created every time a node's start and end values are set.