Refactor code and improve function implementation
The commit improved several class method implementations, particularly in the vector and base classes. A new method for bisection in 2D was added to the bisection.py file. Some class method names were changed for improved clarity, such as get_data to to_data in the api/surface.py file. It also introduced changes aimed to enhance the functionality of the core numeric functions and geometrical computations in the library. The Jax library has been added to pyproject.toml to enhance the computational capabilities of the project.