Add IntersectionImplicitCurve and TraceIntersectionImplicitCurve classes
This commit introduces two new classes: IntersectionImplicitCurve and TraceIntersectionImplicitCurve. The former is designed to calculate the intersection of two 3D implicit surfaces with a given tolerance, while the latter is used to trace intersection curves of implicit surfaces. Moreover, methods for building a KDTree, query points, and rebuilding a tree for efficient intersection calculations have been added.