Refactor numerical methods and geometric calculations
This commit greatly refactors the numerical methods and geometric calculations in the `mmcore` module. The Hess class was relocated, the DEFAULT_H value was widely used as a replacement, the tolerance values were changed and the newton's method parameters were simplified. In the `mmcore/geom/implicit` module, function import locations were updated and calculations were optimized. Lastly, a main execution condition was added for cylinder calculations in the fdm.py file.