Update project to version 0.49.0
1. `pyproject.toml`
Update project version to 0.49.0

Changed the project version from 0.48.1 to 0.49.0 in the `pyproject.toml` file. This change indicates a new release of the project, which may include new features, improvements, or bug fixes. Updating the version number is a standard practice to help with dependency management and version tracking. This update ensures that users and developers are aware of the new changes and can appropriately update their dependencies.