Refactor code and update project configuration
1. mmcore/topo/mesh/triangle/tri.py
Refactor `prepare_args` function for better readability

Removed the usage of `Unpack` from `typing_extensions` and enhanced the `prepare_args` function by using the unpacking (*) operator directly. This change addresses the readability and compatibility of the code by simplifying and modernizing the approach for unpacking boundaries and holes.

2. pyproject.toml
Bump the project version to 0.49.1

Increase the version from 0.49.0 to 0.49.1 to reflect the new changes and updates in the project. This change ensures proper versioning and helps in tracking the updates made to the project.

3. mmcore/numeric/newthon/constrained.py
Add new `constrained` module for Newton's and SQP methods

Introduced a new file `constrained.py` under the `newthon` module which includes multiple functions for Newton's methods and Sequential Quadratic Programming (SQP). These methods are essential for solving constrained optimization problems. The file defines Newton's method for solving systems of equations, computes gradients, Hessians, and implements SQP for optimization. The addition enhances the numerical capabilities of the project.

4. mmcore/compat/step/step_writer.py
Clean up and refactor `step_writer.py` for better maintainability

Removed unused imports, redundant class definitions, and commented-out code, improving the readability and maintainability of the file. Updated the STEP file header to include `mmcore` version information dynamically. Additionally, moved constants to a more appropriate location, ensuring that the codebase is cleaner and follows a more modular approach.